GH-1755 move parser-specific settings out of generic helper package
- where possibly only
