+++ testsuites/benchmark/src/main/java/org/eclipse/rdf4j/benchmark/rio/RDFSizeBenchmarks.java
@@ -36,7 +36,7 @@ import org.eclipse.rdf4j.rio.RDFParser;
 import org.eclipse.rdf4j.rio.RDFWriter;
 import org.eclipse.rdf4j.rio.Rio;
 import org.eclipse.rdf4j.rio.WriterConfig;
-import org.eclipse.rdf4j.rio.helpers.BinaryRDFWriterSettings;
+import org.eclipse.rdf4j.rio.binary.BinaryRDFWriterSettings;
 
 /**
  * <p>
